sports-predictor/requirements.txt
streamlit
sqlmodel
psycopg2-binary
requests
python-dotenv
scipy
